--- yp-core/src/main/java/kr/co/yourplanet/core/entity/portfolio/Studio.java
@@ -1,59 +1,0 @@
-package kr.co.yourplanet.core.entity.portfolio;
-
-import kr.co.yourplanet.core.entity.member.Member;
-import kr.co.yourplanet.core.entity.BasicColumn;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicUpdate;
-
-import javax.persistence.*;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@DynamicUpdate
-@Builder
-public class Studio extends BasicColumn {
-
-    @Id
-    @Column(name = "id")
-    private Long id;
-
-    @OneToOne
-    @MapsId
-    @JoinColumn(name = "id", referencedColumnName = "id")
-    private Member member;
-
-    @Column(name = "description")
-    private String description;
-
-    @Column(name = "toon_name")
-    private String toonName;
-
-    @OneToMany(mappedBy = "studio", fetch = FetchType.LAZY)
-    private List<PortfolioCategoryMap> portfolioCategoryMapList;
-
-    @OneToMany(mappedBy = "studio", fetch = FetchType.LAZY)
-    private List<PortfolioLink> portfolioLinkList;
-
-    public List<String> getCategoryTypes() {
-        return portfolioCategoryMapList.stream()
-                .map(PortfolioCategoryMap::getCategory)
-                .map(Category::getCategoryName).collect(Collectors.toList());
-    }
-
-    public List<String> getPortfolioLinkUrls() {
-        return portfolioLinkList.stream()
-                .map(PortfolioLink::getUrl).collect(Collectors.toList());
-    }
-
-    public void updateStudio(String name, String description) {
-        this.toonName = name;
-        this.description = description;
-    }
-}
