+++ yp-core/src/main/java/kr/co/yourplanet/core/entity/studio/Category.java
@@ -1,0 +1,28 @@
+package kr.co.yourplanet.core.entity.studio;
+
+import kr.co.yourplanet.core.entity.BasicColumn;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.DynamicUpdate;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@DynamicUpdate
+@Builder
+public class Category extends BasicColumn {
+
+    @Id
+    @Column(name = "category_code")
+    private String categoryCode;
+
+    @Column(name = "category_name")
+    private String categoryName;
+}
