+++ yp-core/src/main/java/kr/co/yourplanet/core/entity/studio/PortfolioCategoryMap.java
@@ -1,0 +1,30 @@
+package kr.co.yourplanet.core.entity.studio;
+
+import kr.co.yourplanet.core.entity.BasicColumn;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.DynamicUpdate;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@DynamicUpdate
+@IdClass(PortfolioCategoryMapKey.class)
+public class PortfolioCategoryMap extends BasicColumn {
+
+    @Id
+    @ManyToOne
+    @JoinColumn(name = "studio_id", referencedColumnName = "id")
+    private Studio studio;
+
+    @Id
+    @ManyToOne
+    @JoinColumn(name = "category_code")
+    private Category category;
+}
