+++ yp-core/src/main/java/kr/co/yourplanet/core/entity/studio/PortfolioLinkKey.java
@@ -1,0 +1,13 @@
+package kr.co.yourplanet.core.entity.studio;
+
+import kr.co.yourplanet.core.entity.instagram.InstagramMedia;
+import lombok.Data;
+
+import java.io.Serializable;
+
+@Data
+public class PortfolioLinkKey implements Serializable {
+
+    private Long studio;
+    private String media;
+}
