+++ yp-core/src/main/java/kr/co/yourplanet/core/enums/PostDurationMonthType.java
@@ -1,0 +1,9 @@
+package kr.co.yourplanet.core.enums;
+
+public enum PostDurationMonthType {
+    ONE_MONTH,
+    TWO_MONTH,
+    THREE_MONTH,
+    SIX_MONTH,
+    MORE_THAN_ONE_YEAR
+}
