--- yp-core/src/main/java/kr/co/yourplanet/core/enums/PriceOptionType.java
@@ -1,7 +1,0 @@
-package kr.co.yourplanet.core.enums;
-
-public enum PriceOptionType {
-    DEFAULT,
-    UNPROVIDED,
-    PROVIDED
-}
