+++ yp-core/src/main/java/kr/co/yourplanet/core/enums/ProvisionType.java
@@ -1,0 +1,7 @@
+package kr.co.yourplanet.core.enums;
+
+public enum ProvisionType {
+    DEFAULT,
+    UNPROVIDED,
+    PROVIDED
+}
