--- yp-core/src/main/java/kr/co/yourplanet/core/enums/UploadPeriod.java
@@ -1,9 +1,0 @@
-package kr.co.yourplanet.core.enums;
-
-public enum UploadPeriod {
-    ONE_MONTH,
-    TWO_MONTH,
-    THREE_MONTH,
-    SIX_MONTH,
-    MORE_THAN_ONE_YEAR
-}
