--- yp-online/src/main/java/kr/co/yourplanet/online/business/portfolio/dto/CategoryType.java
@@ -1,5 +1,0 @@
-package kr.co.yourplanet.online.business.portfolio.dto;
-
-public class CategoryType {
-    public String categoryType;
-}
