--- yp-online/src/main/java/kr/co/yourplanet/online/business/portfolio/repository/CategoryRepository.java
@@ -1,10 +1,0 @@
-package kr.co.yourplanet.online.business.portfolio.repository;
-
-import kr.co.yourplanet.core.entity.portfolio.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface CategoryRepository extends JpaRepository<Category, String> {
-    List<Category> findAllByCategoryCodeIn(List<String> categoryCodeList);
-}
