--- yp-online/src/main/java/kr/co/yourplanet/online/business/portfolio/repository/PortfolioLinkRepository.java
@@ -1,11 +1,0 @@
-package kr.co.yourplanet.online.business.portfolio.repository;
-
-import kr.co.yourplanet.core.entity.portfolio.PortfolioLink;
-import kr.co.yourplanet.core.entity.portfolio.Studio;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.time.LocalDateTime;
-
-public interface PortfolioLinkRepository extends JpaRepository<PortfolioLink, Long>{
-    void deleteAllByStudioAndCreateDateBeforeAndUpdateDateBefore(Studio studio, LocalDateTime localDateTime, LocalDateTime localDateTime2);
-}
