--- yp-online/src/main/java/kr/co/yourplanet/online/business/portfolio/repository/StudioRepository.java
@@ -1,10 +1,0 @@
-package kr.co.yourplanet.online.business.portfolio.repository;
-
-import kr.co.yourplanet.core.entity.portfolio.Studio;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface StudioRepository extends JpaRepository<Studio, Long> {
-    Optional<Studio> findByMemberId(Long memberId);
-}
