--- yp-online/src/main/java/kr/co/yourplanet/online/business/portfolio/repository/TempPriceRepository.java
@@ -1,7 +1,0 @@
-package kr.co.yourplanet.online.business.portfolio.repository;
-
-import kr.co.yourplanet.online.business.portfolio.domain.TempPrice;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TempPriceRepository extends JpaRepository<TempPrice, Long> {
-}
