+++ yp-online/src/main/java/kr/co/yourplanet/online/business/studio/repository/InstagramMediaRepository.java
@@ -1,0 +1,7 @@
+package kr.co.yourplanet.online.business.studio.repository;
+
+import kr.co.yourplanet.core.entity.instagram.InstagramMedia;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InstagramMediaRepository extends JpaRepository<InstagramMedia, String> {
+}
