+++ yp-online/src/main/java/kr/co/yourplanet/online/business/studio/repository/StudioRepository.java
@@ -1,0 +1,10 @@
+package kr.co.yourplanet.online.business.studio.repository;
+
+import kr.co.yourplanet.core.entity.studio.Studio;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface StudioRepository extends JpaRepository<Studio, Long> {
+    Optional<Studio> findByMemberId(Long memberId);
+}
