+++ yp-online/src/main/java/kr/co/yourplanet/online/business/studio/repository/TempPriceRepository.java
@@ -1,0 +1,7 @@
+package kr.co.yourplanet.online.business.studio.repository;
+
+import kr.co.yourplanet.core.entity.studio.TempPrice;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TempPriceRepository extends JpaRepository<TempPrice, Long> {
+}
